azure-connection/vsc_azure/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-core==1.35.0
azure-iot-device==2.14.0
azure-iot-hub==2.6.1
blinker==1.9.0
certifi==2025.7.9
charset-normalizer==3.4.2
click==8.2.1
deprecation==2.1.0
Flask==3.1.1
idna==3.10
isodate==0.7.2
itsdangerous==2.2.0
janus==2.0.0
Jinja2==3.1.6
joblib==1.5.1
MarkupSafe==3.0.2
msrest==0.7.1
numpy==2.3.1
nvidia-nccl-cu12==2.27.5
oauthlib==3.3.1
packaging==25.0
paho-mqtt==1.6.1
pandas==2.3.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.4
requests-oauthlib==2.0.0
requests-unixsocket2==1.0.0
scikit-learn==1.7.0
scipy==1.16.0
six==1.17.0
threadpoolctl==3.6.0
typing_extensions==4.14.1
tzdata==2025.2
uamqp==1.6.11
urllib3==2.5.0
Werkzeug==3.1.3
xgboost==3.0.2
